requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
click>=8.0.0

# AI/ML dependencies
scikit-learn>=1.1.0
tensorflow>=2.10.0  # Optional for advanced ML

# Visualization
plotly>=5.0.0
matplotlib>=3.5.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Logging and utilities
structlog>=22.0.0
tqdm>=4.64.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991 